create custom hook for search results
